requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
anthropic==0.40.0
requests==2.32.3
pyyaml==6.0.2
aiosqlite==0.20.0
aiosmtplib==3.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
